Updates event schedule migrations to use bigIncrements for IDs and ensure consistency in data types

+++ database/migrations/2020_06_20_140932_create_event_schedule_event_speaker_table.php
@@ -17,7 +17,7 @@ class CreateEventScheduleEventSpeakerTable extends Migration
             $table->bigIncrements('id');
             $table->bigInteger('event_speaker_id')->unsigned()->nullable();
             $table->foreign('event_speaker_id')->references('id')->on('event_speakers');
-            $table->integer('event_schedule_id')->unsigned()->nullable();
+            $table->bigInteger('event_schedule_id')->unsigned()->nullable();
             $table->foreign('event_schedule_id')->references('id')->on('event_schedules');
             $table->timestamps();
         });
